requirement.txt
# Project: self-media-auto (IP 爆款制造机)
# Generated from pyproject.toml and source code analysis

# Core Dependencies (from pyproject.toml)
browserforge>=1.2.4
curl-cffi>=0.14.0
faster-whisper>=1.2.1
httpx>=0.28.1
loguru>=0.7.3
playwright>=1.58.0
requests>=2.32.5
scrapling>=0.4.2
videofetch>=0.6.8
yt-dlp>=2026.3.13

# Additional Dependencies (identified from source and setup scripts)
python-dotenv          # Supporting load_dotenv
duckduckgo-search      # Supporting duckduckgo_search
xiaohongshu-cli        # Mentioned in setup.ps1
playwright-python      # Base for playwright

# Development / Utility
setuptools             # Supporting -e . (editable install)
pip>=23.0

# System Level Dependencies (Non-Python)
# ffmpeg (Required for faster-whisper/ASR)
# nodejs/npm (Required for baoyu-skills)
